refactor: cleanup panel imports
